% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_bootstrap.R
\name{bootstrap_sampling}
\alias{bootstrap_sampling}
\title{Generate bootstrapped random samples from a dataset}
\usage{
bootstrap_sampling(data, folds)
}
\arguments{
\item{data}{The dataset from which to sample.}

\item{folds}{The number of folds for bootstrapping.}
}
\value{
A list of bootstrapped random samples.
}
\description{
This function generates bootstrapped random samples of a specified size from a dataset
with a specified number of folds.
}
\examples{
# Generate bootstrapped random samples with 5 folds
bootstrap_sampling(letters, folds = 5)

}
