% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_clustered.R
\name{cluster_sampling}
\alias{cluster_sampling}
\title{Two-stage cluster sampling}
\usage{
cluster_sampling(data, clusters, nclusters, within_cluster)
}
\arguments{
\item{data}{The dataset from which to sample.}

\item{clusters}{A vector specifying the size of the clusters in the dataset.
That is, the number of rows corresponding to each cluster. It is assumed the
dataset is ordered by cluster.}

\item{nclusters}{The desired number of clusters to include in the sample.}

\item{within_cluster}{The desired number of elements to sample from within each cluster.}
}
\value{
A random sample of the data
}
\description{
This function does random sampling within clusters. That is, it first selects
clusters at random (without replacement), and then samples from within those
clusters (with replacement).
}
\examples{
# Generate a sample of 5 observations from each of 2 of the 3 clusters
cluster_sampling(letters, clusters = c(7, 6, 13), nclusters = 2,
  within_cluster = 5)
}
