% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_random.R
\name{random_sampling}
\alias{random_sampling}
\title{Generate a simple random sample from a dataset}
\usage{
random_sampling(data, size)
}
\arguments{
\item{data}{The dataset from which to sample.}

\item{size}{The desired size of the random sample.}
}
\value{
A simple random sample of the specified size.
}
\description{
This function generates a simple random sample of a specified size from a dataset.
Each element in the dataset has an equal probability of being chosen.
}
\examples{
# Generate a simple random sample of size 100
sample <- random_sampling(letters, size = 100)
print(sample)
}
