% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_stratified.R
\name{stratified_sampling}
\alias{stratified_sampling}
\title{Generate a stratified random sample from a dataset}
\usage{
stratified_sampling(data, strata, weights, size)
}
\arguments{
\item{data}{The dataset from which to sample.}

\item{strata}{A vector specifying the sizes of the strata within the dataset.
That is, the number of rows corresponding to each stratum. It is assumed the
dataset is ordered by strata.}

\item{weights}{A vector specifying the weights of the strata for sampling.}

\item{size}{The desired size of the stratified random sample.}
}
\value{
A stratified random sample of the specified size.
}
\description{
This function generates a stratified random sample of a specified size from a dataset,
where the strata sizes and weights are provided as input.
}
\examples{
# Generate a stratified random sample of size 100 with given strata sizes
strata <- c(5, 9, 12)
weights <- c(0.5, 0.3, 0.2)
sample <- stratified_sampling(letters, strata, weights, size = 100)
table(sample)

}
