% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_systematic.R
\name{systematic_sampling}
\alias{systematic_sampling}
\title{Generate a systematic random sample from a dataset}
\usage{
systematic_sampling(data, interval)
}
\arguments{
\item{data}{The dataset from which to sample.}

\item{interval}{The sampling interval.}
}
\value{
A systematic random sample
}
\description{
This function generates a systematic random sample from a dataset
with a specified sampling interval.
}
\examples{
# Generate a systematic random sample with an interval of 5
sample <- systematic_sampling(letters, interval = 5)
print(sample)

}
